newapp/requirements.txt
Django==3.2.6
asgiref==3.4.1
autopep8==1.5.7
azure-common==1.1.27
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
django-storages==1.11.1
djangorestframework==3.12.4
docopt==0.6.2
idna==3.2
mssql-django==1.0
Pillow==8.3.1
pipreqs==0.4.10
pycodestyle==2.7.0
pycparser==2.20
pyodbc==4.0.31
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
yarg==0.1.9
